requirements.txt
autogen-agentchat~=0.2
autogen-agentchat[ipython]~=0.2
python-dotenv==0.19.1
firecrawl-py
aiohttp
asyncio
openai>=1.0.0
bs4
flaml[automl]
selenium>=4.0.0
webdriver_manager>=3.8.0
requests>=2.25.1
autogen-agentchat~=0.2



